brain/requirements.txt
python-telegram-bot==20.7
kubernetes==28.1.0
redis==5.0.1
groq==0.9.0